Added state functionality to calculate dates
Several functions were added in this update primarily focusing on
calculating dates for the roadtrip for use in the eventbrite API.
The eventbrite api's framework was added and prints out to console
to verify it's working. Primarily, getLeg function now updates
legDates variable with the dates of the leg. The index corresponds
so legDates[0] is the date of leg[0]. legDates variable was also
initialized.

Additionally, getDates function was created. This function calculates
all dates for the journey and pushes them to the dates array. It bases
it's calculations off of the first date the user inputs and all the
subsequent lengths of stay. Detailed comments are within the getDates
function.