Updated watchFormSubmit to use required and event.preventDefault
WatchFormSubmit function was split into two functions. The first
portion looks to see if destinatino-form is being submitted and then
prevents default. The second portion is for the required tags. If
proceed returns false, the user is notified that they are missing
required data, if it runs true then the application runs.