Began catching user and formatting errors
Fixed display event brite data to display events
event if there is only one event left to view. Previously
it would require at least 2 events in order of the div to close
properly in order for it to display on screen. Added an if statement
to catch when there is only one event.

In addition, rearranged watchSubmit to call functions before
displaying them on screen. In later patches will catch user
inputting bad addresses and asking them to enter a valid one.